Applied Rob's solution
My solution was a single listener method for all the buttons, set with android:onClick in the activity_main.xml, and start the proper media file by "filtering" the button ID with a switch statement. Rob's solution was shorter and I learned that you can access the ID as a String.

I also added an app icon from iconfinder.com